found a way to make the maze Bake with the tree area being unwalkable.
There could be a better fix for this will look into this more tomorrow. Justin if you are reading these notes then if you collapse NavMesh you will see two UnwalkableMesh Objects attached to a script and you can basically do what we did with the blocks in LOTR but some info I was looking at said that there may be a better way to do it through the Terrain Tools.